[BBC-8993] ItemBigData.mdx
