--- src/itemBigData/spec.md
@@ -1,15 +1,0 @@
-# ItemBigData
-
-<!-- STORY -->
-
----
-
-### Normal state
-
-- Used to display a specific and important data.
-- They’re not clickable.
-
-### Long texts
-
-- **Data**: _it can’t drop line_.
-- **Additionnal Info**: there’s no limit on the number of lines.
